squash! Adjust StepsTranslator behaviour
When the requested number of steps passed to convertStepsToDomPoint is
beyond the document end in, or before the start of the document, an
exception will be thrown. This closes a slight hole where asking for a
step beyond the document would return a non-step position. It also makes
the invalid usage of the interface more obvious.